docs(ui): add comments on pull to load implementation
