+++ app/src/main/java/me/ash/reader/ui/page/home/reading/PullToLoad.kt
@@ -1,11 +1,11 @@
 package me.ash.reader.ui.page.home.reading
 
 
-import android.util.Log
 import androidx.compose.animation.core.FloatExponentialDecaySpec
 import androidx.compose.animation.core.animate
 import androidx.compose.animation.core.animateDecay
 import androidx.compose.foundation.MutatorMutex
+import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.State
@@ -15,60 +15,38 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import kotlin.math.abs
-import kotlinx.coroutines.CoroutineScope
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.pullrefresh.pullRefreshIndicatorTransform
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.input.nestedscroll.NestedScrollConnection
 import androidx.compose.ui.input.nestedscroll.NestedScrollSource
 import androidx.compose.ui.input.nestedscroll.NestedScrollSource.Companion.Drag
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.debugInspectorInfo
-import androidx.compose.ui.platform.inspectable
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.Velocity
+import androidx.compose.ui.unit.dp
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.launch
+import kotlin.math.abs
 import kotlin.math.sqrt
 
 private const val TAG = "PullRelease"
 
 /**
- * A nested scroll modifier that provides scroll events to [state].
+ * A [NestedScrollConnection] that provides scroll events to a hoisted [state].
  *
- * Note that this modifier must be added above a scrolling container, such as a lazy column, in
- * order to receive scroll events. For example:
+ * Note that this modifier must be added above a scrolling container using [Modifier.nestedScroll],
+ * such as a lazy column, in order to receive scroll events.
  *
- * @sample androidx.compose.material.samples.PullRefreshSample
+ * And you should manually handle the offset of components
+ * with [PullToLoadState.progress] or [PullToLoadState.offsetFraction]
  *
- * @param state The [PullToLoadState] associated with this pull-to-refresh component.
- * The state will be updated by this modifier.
+ * @param state The [PullToLoadState] associated with this pull-to-load component.
+ * The state will be updated by this connection.
  * @param enabled If not enabled, all scroll delta and fling velocity will be ignored.
+ * @param onScroll Used for detecting if the reader is scrolling down
  */
-// TODO(b/244423199): Move pullRefresh into its own material library similar to material-ripple.
-@ExperimentalMaterialApi
-fun Modifier.pullToLoad(
-    state: PullToLoadState,
-    enabled: Boolean = true,
-    onScroll: (Float) -> Unit
-) = inspectable(inspectorInfo = debugInspectorInfo {
-    name = "pullRefresh"
-    properties["state"] = state
-    properties["enabled"] = enabled
-}) {
-    Modifier.nestedScroll(
-        ReaderNestedScrollConnection(
-            state = state,
-            enabled = enabled,
-            onScroll = onScroll
-        )
-    )
-}
-
 class ReaderNestedScrollConnection(
     private val state: PullToLoadState,
     private val enabled: Boolean,
@@ -113,16 +91,14 @@ class ReaderNestedScrollConnection(
 /**
  * Creates a [PullToLoadState] that is remembered across compositions.
  *
- * Changes to [refreshing] will result in [PullToLoadState] being updated.
+ * Changes from [ReaderNestedScrollConnection] will result in [PullToLoadState] being updated.
  *
- * @sample androidx.compose.material.samples.PullRefreshSample
  *
- * @param refreshing A boolean representing whether a refresh is currently occurring.
- * @param onLoadNext The function to be called to trigger a refresh.
+ * @param key Key used for remembering the state
+ * @param onLoadNext The function to be called to load the next item when pulled up.
+ * @param onLoadPrevious The function to be called to load the previous item when pulled down.
  * @param loadThreshold The threshold below which, if a release
- * occurs, [onLoadNext] will be called.
- * @param refreshingOffset The offset at which the indicator will be drawn while refreshing. This
- * offset corresponds to the position of the bottom of the indicator.
+ * occurs, [onLoadNext] or [onLoadPrevious] will be called.
  */
 @Composable
 @ExperimentalMaterialApi
@@ -130,15 +106,13 @@ fun rememberPullToLoadState(
     key: Any?,
     onLoadPrevious: () -> Unit,
     onLoadNext: () -> Unit,
-    onThresholdReached: () -> Unit,
-    loadThreshold: Dp = PullRefreshDefaults.RefreshThreshold,
+    loadThreshold: Dp = PullToLoadDefaults.LoadThreshold,
 ): PullToLoadState {
-    require(loadThreshold > 0.dp) { "The refresh trigger must be greater than zero!" }
+    require(loadThreshold > 0.dp) { "The load trigger must be greater than zero!" }
 
     val scope = rememberCoroutineScope()
     val onNext = rememberUpdatedState(onLoadNext)
     val onPrevious = rememberUpdatedState(onLoadPrevious)
-    val onLoad = rememberUpdatedState(onThresholdReached)
     val thresholdPx: Float
 
     with(LocalDensity.current) {
@@ -148,7 +122,6 @@ fun rememberPullToLoadState(
     val state = remember(key, scope) {
         PullToLoadState(
             animationScope = scope,
-            onThresholdReached = onLoad,
             onLoadPrevious = onPrevious,
             onLoadNext = onNext,
             threshold = thresholdPx
@@ -163,35 +136,40 @@ fun rememberPullToLoadState(
 }
 
 /**
- * A state object that can be used in conjunction with [pullToLoad] to add pull-to-refresh
+ * A state object that can be used in conjunction with [ReaderNestedScrollConnection] to add pull-to-load
  * behaviour to a scroll component. Based on Android's SwipeRefreshLayout.
  *
  * Provides [progress], a float representing how far the user has pulled as a percentage of the
- * refreshThreshold. Values of one or less indicate that the user has not yet pulled past the
+ * [threshold]. Values of one or less indicate that the user has not yet pulled past the
  * threshold. Values greater than one indicate how far past the threshold the user has pulled.
  *
- * Can be used in conjunction with [pullRefreshIndicatorTransform] to implement Android-like
- * pull-to-refresh behaviour with a custom indicator.
  *
  * Should be created using [rememberPullToLoadState].
  */
 class PullToLoadState internal constructor(
     private val animationScope: CoroutineScope,
-    private val onThresholdReached: State<() -> Unit>,
     private val onLoadPrevious: State<() -> Unit>,
     private val onLoadNext: State<() -> Unit>,
     threshold: Float
 ) {
     /**
-     * A float representing how far the user has pulled as a percentage of the refreshThreshold.
+     * A float representing how far the user has pulled as a percentage of the [threshold].
      *
      * If the component has not been pulled at all, progress is zero. If the pull has reached
      * halfway to the threshold, progress is 0.5f. A value greater than 1 indicates that pull has
-     * gone beyond the refreshThreshold - e.g. a value of 2f indicates that the user has pulled to
-     * two times the refreshThreshold.
+     * gone beyond the [threshold] - e.g. a value of 2f indicates that the user has pulled to
+     * two times the [threshold].
      */
     val progress get() = abs(offsetPulled) / threshold
 
+    /**
+     * The offset fraction calculated from [progress] and [status],
+     * This fraction grows in linear when the [progress] is no greater than 1,
+     * then grows exponentially with the rate 1/2 if the [progress] greater than 1. - e.g. a value
+     * of 2f indicates that the user has pulled to **four** times the [threshold].
+     *
+     * @return The offset fraction currently of this state, could be negative if the content is pulling up
+     */
     val offsetFraction: Float get() = calculateOffsetFraction()
 
     sealed interface Status {
@@ -222,7 +200,6 @@ class PullToLoadState internal constructor(
     private var _threshold by mutableFloatStateOf(threshold)
 
     internal fun onPull(pullDelta: Float): Float {
-        val statusBefore = status
         val consumed = if (offsetPulled.signOpposites(offsetPulled + pullDelta)) {
             -offsetPulled
         } else {
@@ -236,15 +213,10 @@ class PullToLoadState internal constructor(
 
 
         offsetPulled += consumed
-        val statusAfter = status
-        if ((statusAfter is Status.PulledUp || statusAfter is Status.PulledDown) && statusBefore != statusAfter) {
-            onThresholdReached.value()
-        }
         return consumed
     }
 
     internal fun onRelease(velocity: Float): Float {
-        Log.d(TAG, "onPull: $velocity")
 //        val consumed = when {
 //            // We are flinging without having dragged the pull refresh (for example a fling inside
 //            // a list) - don't consume
@@ -256,6 +228,8 @@ class PullToLoadState internal constructor(
 //            else -> velocity
 //        }
         when (status) {
+            // We don't change the pull offset here because the animation for loading another content
+            // should be handled outside, and this state will be soon disposed
             Status.PulledDown -> {
                 onLoadPrevious.value()
             }
@@ -265,6 +239,7 @@ class PullToLoadState internal constructor(
             }
 
             else -> {
+                // Snap to 0f and hide the indicator
                 animateDistanceTo(0f)
             }
         }
@@ -331,10 +306,10 @@ private fun Float.signOpposites(f: Float): Boolean =
  * Default parameter values for [rememberPullToLoadState].
  */
 @ExperimentalMaterialApi
-object PullRefreshDefaults {
+object PullToLoadDefaults {
     /**
-     * If the indicator is below this threshold offset when it is released, a refresh
+     * If the indicator is below this threshold offset when it is released, the load action
      * will be triggered.
      */
-    val RefreshThreshold = 120.dp
+    val LoadThreshold = 120.dp
 }
